msm8998-common: overlay: fwb: reduce networkTransitionTimeout
Devices nowadays are very fast in switching networks,
most of the times they even go unnoticed.

In a case like in Germany where the network infrastructure
isn't as good, the transitions sometimes take very long
or just never happen (= timeout is triggered) so I'm reducing
this to improve network transitioning in such cases.

Change-Id: Ic2a37fbadbdf20af8de1c8bfe1f3ff62acfb5cf4
Signed-off-by: Simao Gomes Viana <xdevs23@outlook.com>
Signed-off-by: NurKeinNeid <mralexman3000@gmail.com>